fix(logging): emit server start success message only after server boot
Previously, the "Server started listening on ..." message would appear before the server came up. In
case the server could not come up, e.g. because it's HTTPS and it doens't have the certificates
availble, this would show in the logs like the server is healthy which in fact it wasn't.